==> emergency_department_encounters_project/03 Scripts/0.Part1_Dataset_Conversion.ipynb <==
### This script contains the following sections:

#### 1. Importing libraries
#### 2. File Conversion
#### 3. Column Section for Future Analysis

# 1. Imports

import pandas as pd
import numpy as np
import os
import glob

path = r'C:\Users\Achievement 6\data_emergency_ex6.1\Data'

# 2. File Conversion

#import multiple excel files, extract data sheet, save as csv files.

files = glob.glob(path + '\Original\1.Raw\*.xlsx')
df_list = []
df_full_list = []
for excel_file in files:
    
    tmp = excel_file.split('.')[1] # remove the '.xlsx' file type at the end of file name
    file = tmp.split('\\')[4]+'.csv' # split file path and add .csv
    
    df = pd.read_excel(os.path.join(path,'Original','1.Raw',excel_file), sheet_name='Data')
    df.to_csv(os.path.join(path,'Original','Converted',file), index = 0)
    
    df_full_list.append(file)
    df_list.append(file[:4])
    
    print(file)

print(df_full_list)
# print(df_list)

# 3. Column Selection

#import converted data files

imported_df_list = []

for i in range(len(df_full_list)):
        temp_df = pd.read_csv(os.path.join(path,'Original','2.Converted', df_full_list[i]))
                              
        imported_df_list.append(temp_df)
                                   
print(len(imported_df_list))   

# Compare list of columns in the datasets

df_cols = list()
for df in imported_df_list:
    df_cols.append(df.columns)

print(df_cols[0])

# add mising columns to earlier datasets

mis_col = ['MSSA_DESIGNATION','RURAL_HOSPITAL_FLAG', 'SelfPay']
mis_col2 = ['TRAUMA_CENTER_DESC']

for count,df in enumerate(imported_df_list[0:-8]):
    df[mis_col[0]] = np.nan
    df[mis_col[1]] = np.nan
    df[mis_col[2]] = np.nan
    df.to_csv(os.path.join(path,'Original','2.Converted', str(df_full_list[count])))
#     print(df_full_list[count])
    

# imported_df_list[6].columns

# Add self pay to 2012 and 2013 datasets

for count,df in enumerate(imported_df_list[7:9]):
    count += 7
    df[['SelfPay']] = np.nan
    df.to_csv(os.path.join(path,'Original','2.Converted', str(df_full_list[count])))

# imported_df_list[7:9]

# Add missing language columns for 2005 to 2009
lang_cols = ['Spanish', 'All_Other', 'English']

for count,df in enumerate(imported_df_list[0:4]):
    df[lang_cols[0]] = np.nan
    df[lang_cols[1]] = np.nan
    df[lang_cols[2]] = np.nan
    df.to_csv(os.path.join(path,'Original','2.Converted', str(df_full_list[count])))
    print(count)

# add missing column related about trauma_ctr to the 2016:2018 dataset

for count, df in enumerate(imported_df_list[11:14]):
    count += 11
    df[mis_col2[0]] = np.nan
    df.to_csv(os.path.join(path,'Original','2.Converted', str(df_full_list[count])))
    print("Saved {}".format(str(df_full_list[count])))
