==> emergency_department_encounters_project/03 Scripts/1.Part1_Data_Concat.ipynb <==
### This script contains the following sections:

#### 1. Importing libraries
#### 2. Import and Convert Dataset for Variable Selection
#### 3. Import and Combine Converted Datasets

# 1. Library Imports

import pandas as pd
import numpy as np
import os
import glob

path = r'C:\Users\Achievement 6\data_emergency_ex6.1\Data'
col_path = r'C:\Users\Achievement 6\EX6.1'

# 2. Imported Converted Datasets for Variable Selection

files = glob.glob(path + '\Original\[2].Converted\*.csv')
# files

#import multiple csv file

df_list = {}

for file in files:
  
    file_name = file.split('\\')[11]
    
    file_df = pd.read_csv(os.path.join(path,'Original','2.Converted', file_name), index_col = 0)

    file_name = 'df_' + file_name.split('-')[0]
    
    df_list[file_name] = file_df.reset_index()
    
#     print(files)

print(df_list.keys())

# insert year column for datasets without 
col = 'datayear'

for df_name, df in df_list.items():
    if col not in df.columns:
        df[col] = np.nan
        df_name = df_name.strip()[3:] + '-hospital-emergency-department-characteristics-by-facility-pivot-profile.csv'
        df.to_csv(os.path.join(path,'Original','2.Converted', df_name))
        print(df_name, df.columns[-1])
    

#change column names for all datasets before concat
#Since the column names vary, I chose to ue the list of column names from the 2020 dataset. 

#load chosen columns for each dataset. The columns were chosen manually due to inconsistencies in the 
# column names and their order in the individual datasets

cols_names_ordered = pd.read_csv(os.path.join(col_path,'column_names.csv'), header = None)

cols_names_ordered = cols_names_ordered.iloc[:,1:]
# cols_names.isna().sum()
cols_names_ordered

# x = ['fac_id','COUNTY']
# x = cols_names_ordered.iloc[0,:].tolist()
# x = cols_names_ordered[['fac_id','COUNTY']]
# df_list['df_2010'].columns.str.replace(' ','')

# cols_names_ordered.iloc[,0]
# df_list['df_2012']


df_list['df_2013']

# rename all datasets before column selection
#make a new df list with df with chosen columns

df_list_updated = {}
i = 0 #declared I inside and it was not looping. wow! took a while
for df_name, df in df_list.items():
    df_filtered = df.copy()
    df_filtered = df_filtered[cols_names_ordered.iloc[i]]
    df_list_updated[df_name] = df_filtered
    print(df_name, df_filtered.shape, 'index for df is {}'.format(i))
    i = i + 1

# df_list_updated['df_2005'].iloc[:,0]
df_list_updated['df_2005'].columns[0]

#Selected chosen columns excluding the first one (2020 column names). This will be applied to all datasets

col_names_selected = cols_names_ordered.iloc[10,].tolist()
print(col_names_selected)

# give the columns a unified names across all datasets

for df_name, df in df_list_updated.items():
    
    for i,col in enumerate(df.columns):
        df.rename(columns={col: col_names_selected[i]}, inplace = True)
#         print(i, col_names_selected[i])

    df['datayear'] = df_name.strip()[3:] # fill missing year values before merging dataset
    
    df_name = 'cols_standardized' + df_name.strip()[3:] + '-hospital-emergency-department-characteristics-by-facility-pivot-profile.csv'
    
    df.to_csv(os.path.join(path,'Original','3.Updated', df_name), index = 0)
    
    print(df_name, 'was saved successfully')
    
#     df_name = 'cols_standardized' + df_name.strip()[3:] + '-hospital-emergency-department-characteristics-by-facility-pivot-profile.csv'
#     df.to_csv(os.path.join(path,'Original','Converted', df_name))
    

# 3. Load and Combine Converted Datasets

files = glob.glob(path + '/Original/[3].Updated/*.csv')

prepared_df = {}

for file in files:
  
    file_name = file.split('\\')[10]
    
    file_df = pd.read_csv(os.path.join(path,'Original','3.Updated', file_name), index_col = 0)

    file_name = file_name.split('-')[0]
    
    prepared_df[file_name] = file_df.reset_index()

prepared_df.keys()

# prepared_df.values()

df_to_merge = prepared_df.values()
df_combined = pd.concat(df_to_merge).drop_duplicates().reset_index(drop = True)
df_combined.shape

df_combined.tail()

df_combined.to_csv(os.path.join(path,'Prepared', 'df_combined.csv'), index = 0)

df_combined.isnull().sum()
