==> emergency_department_encounters_project/03 Scripts/2.Part2_Data_Cleaning.ipynb <==
### This script contains the following sections:

#### 1. Importing libraries and Dataset
#### 2. Mixed-type Variable Check and Fixes
#### 3. Missing Value Checks and Fixes

# 1. Imports

import pandas as pd
import numpy as np
import os
# import glob

data_path = r'C:\Users\\Data'
res_path = r'C:\Users\Abu\\data_emergency_ex6.1'

#load combined dataset
df = pd.read_csv(os.path.join(data_path, 'Prepared','df_combined.csv'))
df.head()

df.shape

# 2. Mixed-type Variable Check and Fixes

df.describe()

# change column types to categorical

df[['datayear', 'oshpd_id2']] = df[['datayear', 'oshpd_id2']].astype('str')
df.info()

# cols = ['datayear', 'oshpd_id2']
# for col in cols:
#     df[col] = df[col].astype('str')

# Check for mixed variable types within columns

mixed_list = []

for col in df.columns:
    mixed_type = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
    if len(df[mixed_type]) > 0:
        print(col)
        mixed_list.append(col)

# change column type for all mixed type variables

for col in mixed_list:
    df[col] = df[col].astype('str')

df['LICENSE_CATEGORY_DESC'].value_counts(dropna = False)

# for val in df['LICENSE_CATEGORY_DESC']:
#     print (type(val)) #floats found

df.loc[df['LICENSE_CATEGORY_DESC'] == "General Acute Care Hospital",
       'LICENSE_CATEGORY_DESC'] = 'General Acute Care'
    
df.loc[df['LICENSE_CATEGORY_DESC'] == "Acute Psychiatric Hospital", 
       'LICENSE_CATEGORY_DESC'] = 'Acute Psychiatric'

df.loc[df['LICENSE_CATEGORY_DESC'] == "Chemical Dep. Recovery Hospital", 
       'LICENSE_CATEGORY_DESC'] = 'Chemical Dependency Recovery'

df.loc[df['LICENSE_CATEGORY_DESC'] == "Chemical Dependency Recovery Hospital", 
       'LICENSE_CATEGORY_DESC'] = 'Chemical Dependency Recovery'

df.loc[df['LICENSE_CATEGORY_DESC'] == "nan", 
       'LICENSE_CATEGORY_DESC'] = np.nan


df['LICENSE_CATEGORY_DESC'].value_counts(dropna = False)

df['Control_Type_Desc'].value_counts(dropna = False)

# Interesting column, but will drop for now.
df.drop(columns = ['Control_Type_Desc'], inplace = True)
df.shape

df['TRAUMA_CENTER_DESC'].value_counts(dropna = False)

df.drop(columns = ['TRAUMA_CENTER_DESC'], inplace = True)
df.shape

df['ER_SERVICE_LEVEL_DESC'].value_counts(dropna = False)

#Standardize value names for ER servie level description

df.loc[df['ER_SERVICE_LEVEL_DESC'] == 'Emergency - Basic', 'ER_SERVICE_LEVEL_DESC'] = "Basic"

df.loc[df['ER_SERVICE_LEVEL_DESC'] == 'Emergency - Standby', 'ER_SERVICE_LEVEL_DESC'] = "Standby"

df.loc[df['ER_SERVICE_LEVEL_DESC'] == 'Emergency - Comprehensive', 'ER_SERVICE_LEVEL_DESC'] = "Comprehensive"

df.loc[df['ER_SERVICE_LEVEL_DESC'] == '0', 'ER_SERVICE_LEVEL_DESC'] = "Unspecified"

df.loc[df['ER_SERVICE_LEVEL_DESC'] == 'nan', 'ER_SERVICE_LEVEL_DESC'] = "Unspecified"

df['ER_SERVICE_LEVEL_DESC'].value_counts(dropna = False)

df['RURAL_HOSPITAL_DESC'].value_counts(dropna = False)

df.drop(columns = ['RURAL_HOSPITAL_DESC'], inplace = True)
df.shape

df['TEACHING_HOSPITAL_DESC'].value_counts(dropna = False)

#replace values for teaching hospital description using 1/0
#Assumption: the na values for recent datasets indicated that the facility is not a teaching hospital
n = ['NO','N','No']
y = ['YES','Y','Teaching']

for val in df['TEACHING_HOSPITAL_DESC']:
    if (val in n):
        df.loc[df['TEACHING_HOSPITAL_DESC'] == val, 'TEACHING_HOSPITAL_DESC'] = "0"
    elif val in y:
        df.loc[df['TEACHING_HOSPITAL_DESC'] == val, 'TEACHING_HOSPITAL_DESC'] = "1"
#     else:
#         df.loc[df['TEACHING_HOSPITAL_DESC'] == val, 'TEACHING_HOSPITAL_DESC'] = np.nan

df['TEACHING_HOSPITAL_DESC'].value_counts(dropna = False)

dict(df['LICENSED_BED_SIZE'].value_counts(dropna = False))

#Since there are mixed type variables, the column cannot be converted to a string. The code below seperated the string that
# that can be converted to floats, then the floats are converted to a categorical string.

prev_beds = ['Beds 1-49','Beds 50-99','Beds 100-149','Beds 150-199','Beds 200-299',
             'Beds 300-499','Beds 500+']
bed_cat = ['1-49','50-99','100-149','150-199','200-299','300-499','500+']

s = 'Beds'
t = '-'
u = '+'

for val in df['LICENSED_BED_SIZE']:
        if s in str(val):
            if val == prev_beds[0] :
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[0]
                print(val)
            elif val == prev_beds[1]:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[1]
                print(val)
            elif val == prev_beds[2]:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[2]
                print(val)
            elif val == prev_beds[3]:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[3]
                print(val)
            elif val == prev_beds[4]:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[4]
                print(val)
            elif val == prev_beds[5]:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[5]
                print(val)
            elif val == prev_beds[6]:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[6]
                print(val)
        elif t in str(val):
            pass
        elif u in str(val):
            pass
        else:
            if (float(val) in range(1,50)):
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[0]
                print(val)
            elif float(val) in range(50,100):
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[1]
                print(val)
            elif float(val) in range(100,150):
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[2]
                print(val)
            elif float(val) in range(150,200):
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[3]
                print(val)
            elif float(val) in range(200,300):
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[4]
                print(val)
            elif float(val) in range(300,500):
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[5]
                print(val)
            elif float(val) > 500:
                df.loc[df['LICENSED_BED_SIZE'] == val,'LICENSED_BED_SIZE'] = bed_cat[6]
                print(val)
            

df.loc[df['LICENSED_BED_SIZE'].isna(),'LICENSED_BED_SIZE'] = np.nan

df['LICENSED_BED_SIZE'].value_counts(dropna = False)

df['MSSA_DESIGNATION'].value_counts(dropna = False)

df.loc[df['MSSA_DESIGNATION'].isnull(),'MSSA_DESIGNATION'] = np.nan

df['MSSA_DESIGNATION'].value_counts(dropna = False)

# Strip +4 zipcode since the 9 digit zip codes are not required for this analysis.e.g. 12345-6789 -> 12345

count = 0
for val in df['DBA_ZIP_CODE']:
    if len(val) > 5:
        df.loc[df['DBA_ZIP_CODE'] == val,'DBA_ZIP_CODE'] = val[:-5]
        count += 1

print(count, 'zip-codes changed')

#change the size of column names for consistency
for col in df.columns:
    df.rename(columns = {col:col.lower()}, inplace = True)

df.head()

#save updated dataset
df.to_csv(os.path.join(data_path,'Prepared', '1.data_mix_check.csv'), index = 0)

# 3. Dealing with Missing Values

#load previously save dataset. USe excel to change'nan' to blank

df_updated = pd.read_csv(os.path.join(data_path,'Prepared', '1.data_mix_check.csv'))
df_updated.tail()

#For license category description, imputed missing value for one of the facilities in excel by searching for similar id. 
#Will delete the remaining two values in jupyter

dict(df_updated.isnull().sum())

# Dropping NA in categorical columns that are central to my curiostity

df_updated.dropna(axis = 0, how = 'any', subset = ['er_service_level_desc', 'license_category_desc',
                                                   'licensed_bed_size','sex_male'], inplace = True)

#remove missing values in license_category_desc column = 0
df_updated = df_updated[df_updated['licensed_bed_size'] != '0']

df_updated['licensed_bed_size'].value_counts(dropna = False)

#replace na values for teaching hospitals based on assumption, along with other numeric columns where the likely hood
# having a value of zero is high

df_updated[['teaching_hospital_desc','sex_unk','age_unknown']] = df_updated[['teaching_hospital_desc',
                                                                            'sex_unk','age_unknown']].fillna(0)
# df_updated['teaching_hospital_desc'].value_counts(dropna = False)

#view updated dataset
df_updated.head()

dict(df_updated.isnull().sum())

df_updated.to_csv(os.path.join(data_path,'Prepared', '2.miss_val_check.csv'), index = 0)
