==> emergency_department_encounters_project/03 Scripts/3.Additional_Data_Joins_plus_more_wrangling.ipynb <==
### This script contains the following sections:

#### 1. Importing libraries and data 
#### 2. Integrity checks
#### 3. Addtional Data Joins
#### 4. Additional Data Cleaning

# 1. Imports

import pandas as pd
import numpy as np
import os

path = r'C:\Users\data_emergency_ex6.1\Data'

#import datasets to join

df_main = pd.read_csv(os.path.join(path,'Prepared','2.miss_val_check.csv'))

df_main.head()

# df_main['licensed_bed_size'].value_counts(dropna = False)

df_add = pd.read_csv(os.path.join(path,'Original','1.Raw','encounters-by-facility.csv'))

df_add.head()

print('Main df'+ str(df_main.shape), 'supple_df' + str(df_add.shape))

# 2. Integrity Checks

# check the number of unique ids in each dataset

print(df_main.groupby(['oshpd_id2'])['oshpd_id2'].unique().count(),
      df_add.groupby(['oshpd_id'])['oshpd_id'].unique().count())

# Check for na in addtional dataset
df_add.isna().sum()

df_add['er_service_level_desc'] = df_add['er_service_level_desc'].fillna('UNKNOWN')

# df_add['er_service_level_desc'] = df_add['er_service_level_desc'].str.lower()

df_add.isna().sum()

# 3. Additional Data Joins

#since the supplementary dataset has shorter time periods, the supple dataset will become the main dataset.

df_combined = df_add.merge(df_main, left_on = ['oshpd_id','year'], right_on = ['oshpd_id2','datayear'],
             how = 'inner', indicator=True)
df_combined['_merge'].value_counts(dropna = False)

comp = np.where(df_combined['er_service_level_desc_x'].str.lower() == df_combined['er_service_level_desc_y'].str.lower(), 
                True, False)
df_combined['same'] = comp

#Inconsistencies found in the the two datasets. Mainly, varying levels assigned to the same hospitals in the same year
#E.g. Basic and Standby

#Drop columns that are not relevant or redundant the dataset

df_combined.drop(columns = ['oshpd_id2','county_name_y','er_service_level_desc_y','_merge','same','ed_visit','datayear'],
                 inplace = True)

df_combined.rename(columns = {'county_name_x': 'county_name',
                              'type': 'encounter_type',
                              'count': 'encounter_type_total',
                              'er_service_level_desc_x': 'er_service_level'}, inplace = True)

df_combined.columns

df_codes = pd.read_csv((os.path.join(path,'Original','1.Raw','county_codes.csv')))
df_codes.shape

df_codes.head()

df_codes['county_name'] = df_codes['county_name'].str.upper()

# add county codes to the list
df_combined_county = df_combined.merge(df_codes, on = 'county_name', how = 'left', indicator = True)
df_combined_county.head()

df_combined_county['_merge'].value_counts(dropna = False)

print(df_combined_county.columns.tolist())

# Change order of columns and also drop "_merge" and "datayear" column 

df_combined_county = df_combined_county[['year','oshpd_id','county_number', 'county_name', 'facility_name',
                             'dba_address1', 'dba_city', 'dba_zip_code', 
                             'license_category_desc', 'teaching_hospital_desc', 'licensed_bed_size',
                             'mssa_designation','er_service_level', 'encounter_type', 'encounter_type_total', 'sex_male', 'sex_female', 'sex_unk', 'age_01_09', 'age_10_19', 
                             'age_20_29', 'age_30_39', 'age_40_49', 'age_50_59', 'age_60_69', 'age_70_79', 'age_80_',
                             'age_under_1', 'age_unknown', 'asian_pacific_island', 'black', 'hispanic',
                             'native_american_eski', 'other_race', 'unknown_race', 'white', 'acute_care', 
                             'against_medical_advice', 'died', 'hospice_care', 'home_health_service', 'residential_care', 
                             'routine', 'psychiatric_care', 'sn_ic_care', 'medi_cal', 'medicare', 'other_payer',
                             'private_coverage', 'selfpay', 'all_other', 'english', 'spanish']]

# 4. Data Cleaning

# Impute numerical columns with 0

df_combined_county.loc[:,'sex_unk':'spanish'] = df_combined_county.loc[:,'sex_unk':'spanish'].fillna(0)

df_combined_county.head()

#seperate columns from the target pivot columns

piv_cols = [col for col in df_combined_county.columns if col not in ['encounter_type', 'encounter_type_total']]
# piv_cols

# Pivot 'er_service_level', 'encounter_type' columns to make two columns for the former with values from the latter

# year 2012 was removed when pivot_table was used. 
df_combined_pivoted = df_combined_county.pivot(index = piv_cols,
                                columns ='encounter_type', values='encounter_type_total').reset_index()
# df_combined_pivoted.columns.name = None

df_combined_pivoted.columns = df_combined_pivoted.columns.str.lower()

df_combined_pivoted

df_combined_pivoted.shape

# Check for duplicate values

df_combined_pivoted[df_combined_pivoted.duplicated()]

df_combined_pivoted.dropna(axis = 0, how = 'any', subset = ['ed_admit'], inplace = True)
df_combined_pivoted.drop(columns = ['mssa_designation'], inplace = True)

df_combined_pivoted.isna().sum()

df_combined_pivoted['county_name'] = df_combined_pivoted['county_name'].str.title()

df_combined_pivoted['county_name'].unique()

df_combined_pivoted['dba_zip_code'] = df_combined_pivoted['dba_zip_code'].astype('str')

df_combined_pivoted.info()

# df_combined_pivoted.to_csv(os.path.join(path,'Prepared','3.data_join_facilities.csv'), index = 0)
